+++ baltic/__init__.py
@@ -1,1 +1,7 @@
-from .baltic import *
+from .tree import Tree
+from .node import Node
+from .leaf import Leaf
+from .reticulation import Reticulation
+from .clade import Clade
+from .io import *
+from .bt_utils import *
